+++ src/enclave/test/http.cpp
@@ -1,9 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
-#include "../http.h"
-
-#include "../http_builder.h"
+#include "../httpbuilder.h"
+#include "../httpparser.h"
 
 #include <doctest/doctest.h>
 #include <queue>
